feature: improve logging for the Superkey worker
Many of the current logs provided way too much information which made
monitoring the service really hard. The goal of these changes is to
provide that extra information only when an error occurs or when the
"debug" log level is enabled.

For the rest of the informational logs, we are fine with just following
the status of the worker.

RHCLOUD-36959