fix: API Client is unable to contact Sources
Apparently the "URL.String()" method escapes the special characters from
the paths by default, which makes the calls to fetch internal
authentications to fail, because the URL contains brackets.

Also, we were attempting to read the body after it was closed in the
retry attempts, which was causing errors too. By moving it to the
retries loop we fulfill two purposes:

1. Draining the body completely for a new retry and connection reuse.
2. Reading the body to be able to be used in log statements or to be
   marshalled.

RHCLOUD-36959