revert: REST API Client refactor
The refactor intended to be something that would help simplifying the
code and making it easier to read and easier to follow. However, I have
been feeling more and more uneasy about pushing such changes without any
tests, and I do not want these changes to hit production without proper
testing of the refactored client.

Therefore, since the old client simply worked and has proven to work
throughout this time, I am reverting the changes so that we keep using
the "proven-to-work" code, even if I would really like to have a better
and easier-to-follow and use REST API client for the Superkey worker.

Instead, I will leave the suggested changes in a separate PR, and I will
be progressively adding tests until the refactored client is completed.